Call the internal numeric comparator cmp_numerics
Instead of calling the numeric_cmp function through DirectFunctionCall2,
call the internal comparator function to eliminate the overhead (no need
to call InitFunctionCallInfoData, etc.).

This requires either changes to the Postgres (exposing the numeric_cmp,
which is currently static in numeric.c) or copying a lot of code from
numeric.c into the extension.